Revert "Vectors/deprecate legacy introduce named vectors only syntax (#1659)"
This reverts commit f8d7d2d3c243ba0d675dc2fe32f4c54a0b11c691.
