+++ .github/workflows/main.yaml
@@ -20,12 +20,12 @@ concurrency:
 env:
   WEAVIATE_125: 1.25.34
   WEAVIATE_126: 1.26.17
-  WEAVIATE_127: 1.27.27
-  WEAVIATE_128: 1.28.16
-  WEAVIATE_129: 1.29.9
-  WEAVIATE_130: 1.30.11
+  WEAVIATE_127: 1.27.14
+  WEAVIATE_128: 1.28.8
+  WEAVIATE_129: 1.29.8
+  WEAVIATE_130: 1.30.9
   WEAVIATE_131: 1.31.5
-  WEAVIATE_132: 1.32.0-rc.1-098c594
+  WEAVIATE_132: 1.32.0-rc.0-6eb89d6.amd64
 
 jobs:
   lint-and-format:
