+++ .pre-commit-config.yaml
@@ -49,4 +49,4 @@ repos:
       language: node
       pass_filenames: false
       types: [python]
-      additional_dependencies: [pyright@1.1.400]
+      additional_dependencies: [pyright@1.1.398]
