+++ docs/conf.py
@@ -6,6 +6,7 @@
 import contextlib
 import os
 import sys
+
 from importlib.metadata import PackageNotFoundError
 from importlib.metadata import version as version_func
