+++ integration/test_collection.py
@@ -7,11 +7,7 @@ from typing import Any, Callable, Dict, List, Optional, Sequence, TypedDict, Uni
 
 import pytest
 
-from integration.conftest import (
-    CollectionFactory,
-    CollectionFactoryGet,
-    _sanitize_collection_name,
-)
+from integration.conftest import CollectionFactory, CollectionFactoryGet, _sanitize_collection_name
 from integration.constants import WEAVIATE_LOGO_NEW_ENCODED, WEAVIATE_LOGO_OLD_ENCODED
 from weaviate.collections.classes.batch import ErrorObject
 from weaviate.collections.classes.config import (
@@ -44,11 +40,7 @@ from weaviate.collections.classes.internal import (
     ReferenceToMulti,
     _CrossReference,
 )
-from weaviate.collections.classes.types import (
-    PhoneNumber,
-    WeaviateProperties,
-    _PhoneNumber,
-)
+from weaviate.collections.classes.types import PhoneNumber, WeaviateProperties, _PhoneNumber
 from weaviate.exceptions import (
     UnexpectedStatusCodeError,
     WeaviateInsertInvalidPropertyError,
@@ -79,7 +71,7 @@ def test_insert_with_typed_dict_generic(
 
     dummy = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection = collection_factory_get(dummy.name, TestInsert)
     uuid = collection.data.insert(properties=TestInsert(name="some name"))
@@ -95,7 +87,7 @@ def test_insert_with_dict_generic(
 ) -> None:
     dummy = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection = collection_factory_get(dummy.name, Dict[str, str])
     uuid = collection.data.insert(properties={"name": "some name"})
@@ -108,7 +100,7 @@ def test_insert_with_dict_generic(
 def test_insert_with_no_generic(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid = collection.data.insert(properties={"name": "some name"})
     objects = collection.query.fetch_objects()
@@ -120,7 +112,7 @@ def test_insert_with_no_generic(collection_factory: CollectionFactory) -> None:
 def test_insert_with_consistency_level(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     ).with_consistency_level(ConsistencyLevel.ALL)
     uuid = collection.data.insert(properties={"name": "some name"})
     objects = collection.query.fetch_objects()
@@ -142,7 +134,7 @@ def test_insert_bad_input(collection_factory: CollectionFactory) -> None:
 
 def test_delete_by_id(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     uuid = collection.data.insert(properties={})
@@ -153,7 +145,7 @@ def test_delete_by_id(collection_factory: CollectionFactory) -> None:
 
 def test_delete_by_id_consistency_level(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     ).with_consistency_level(ConsistencyLevel.ALL)
 
     uuid = collection.data.insert(properties={})
@@ -232,7 +224,7 @@ def test_insert_many(
 ) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     if not should_error:
         ret = collection.data.insert_many(objects)
@@ -266,7 +258,7 @@ def test_insert_many_all_error(
 ) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         multi_tenancy_config=Configure.multi_tenancy(True),
     )
     with pytest.raises(WeaviateInsertManyAllFailedError) as e:
@@ -286,7 +278,7 @@ def test_insert_many_with_typed_dict(
 
     dummy = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection = collection_factory_get(dummy.name, TestInsertManyWithTypedDict)
     ret = collection.data.insert_many(
@@ -305,7 +297,7 @@ def test_insert_many_with_typed_dict(
 
 def test_insert_many_with_refs(collection_factory: CollectionFactory) -> None:
     ref_collection = collection_factory(
-        name="target", vector_config=Configure.Vectors.self_provided()
+        name="target", vectorizer_config=Configure.Vectorizer.none()
     )
     uuid_to1 = ref_collection.data.insert(properties={})
     uuid_to2 = ref_collection.data.insert(properties={})
@@ -314,7 +306,7 @@ def test_insert_many_with_refs(collection_factory: CollectionFactory) -> None:
         name="source",
         properties=[Property(name="Name", data_type=DataType.TEXT)],
         references=[ReferenceProperty(name="ref_single", target_collection=ref_collection.name)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection.config.add_reference(
         ReferenceProperty.MultiTarget(
@@ -385,7 +377,7 @@ def test_insert_many_with_refs(collection_factory: CollectionFactory) -> None:
 def test_insert_many_error(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     ret = collection.data.insert_many(
         [
@@ -411,7 +403,7 @@ def test_insert_many_error(collection_factory: CollectionFactory) -> None:
 def test_replace(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid = collection.data.insert(properties={"name": "some name"})
     collection.data.replace(properties={"name": "other name"}, uuid=uuid)
@@ -421,7 +413,7 @@ def test_replace(collection_factory: CollectionFactory) -> None:
 @pytest.mark.parametrize("to_uuids", [UUID3, [UUID3]])
 def test_replace_with_refs(collection_factory: CollectionFactory, to_uuids: UUIDS) -> None:
     ref_collection = collection_factory(
-        name="target", vector_config=Configure.Vectors.self_provided()
+        name="target", vectorizer_config=Configure.Vectorizer.none()
     )
     ref_collection.data.insert(properties={}, uuid=UUID1)
     ref_collection.data.insert(properties={}, uuid=UUID2)
@@ -431,7 +423,7 @@ def test_replace_with_refs(collection_factory: CollectionFactory, to_uuids: UUID
         name="source",
         properties=[Property(name="Name", data_type=DataType.TEXT)],
         references=[ReferenceProperty(name="ref", target_collection=ref_collection.name)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid = collection.data.insert(
         properties={"name": "some name"},
@@ -461,7 +453,7 @@ def test_replace_with_refs(collection_factory: CollectionFactory, to_uuids: UUID
 def test_replace_overwrites_vector(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid = collection.data.insert(properties={"name": "some name"}, vector=[1, 2, 3])
     obj = collection.query.fetch_object_by_id(uuid, include_vector=True)
@@ -471,10 +463,7 @@ def test_replace_overwrites_vector(collection_factory: CollectionFactory) -> Non
     collection.data.replace(properties={"name": "other name"}, uuid=uuid)
     obj = collection.query.fetch_object_by_id(uuid, include_vector=True)
     assert obj.properties["name"] == "other name"
-    if collection._connection._weaviate_version.is_lower_than(1, 27, 0):
-        assert len(obj.vector["default"]) == 0
-    else:
-        assert "default" not in obj.vector
+    assert "default" not in obj.vector
 
     collection.data.replace(properties={"name": "real name"}, uuid=uuid, vector=[2, 3, 4])
     obj = collection.query.fetch_object_by_id(uuid, include_vector=True)
@@ -485,7 +474,7 @@ def test_replace_overwrites_vector(collection_factory: CollectionFactory) -> Non
 @pytest.mark.parametrize("to_uuids", [UUID3, [UUID3]])
 def test_update_with_refs(collection_factory: CollectionFactory, to_uuids: UUIDS) -> None:
     ref_collection = collection_factory(
-        name="target", vector_config=Configure.Vectors.self_provided()
+        name="target", vectorizer_config=Configure.Vectorizer.none()
     )
     ref_collection.data.insert(properties={}, uuid=UUID1)
     ref_collection.data.insert(properties={}, uuid=UUID2)
@@ -495,7 +484,7 @@ def test_update_with_refs(collection_factory: CollectionFactory, to_uuids: UUIDS
         name="source",
         properties=[Property(name="Name", data_type=DataType.TEXT)],
         references=[ReferenceProperty(name="ref", target_collection=ref_collection.name)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid = collection.data.insert(
         properties={"name": "some name"},
@@ -539,7 +528,7 @@ def test_types(collection_factory: CollectionFactory, data_type: DataType, value
     name = "name"
     collection = collection_factory(
         properties=[Property(name=name, data_type=data_type)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     uuid_object = collection.data.insert(properties={name: value})
 
@@ -556,7 +545,7 @@ def test_types(collection_factory: CollectionFactory, data_type: DataType, value
 def test_bm25(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     res = collection.data.insert_many(
@@ -573,7 +562,7 @@ def test_bm25(collection_factory: CollectionFactory) -> None:
 def test_bm25_group_by(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     res = collection.data.insert_many(
@@ -601,7 +590,7 @@ def test_bm25_group_by(collection_factory: CollectionFactory) -> None:
 def test_bm25_limit(collection_factory: CollectionFactory, limit: int) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT, tokenization=Tokenization.WORD)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     res = collection.data.insert_many(
@@ -619,7 +608,7 @@ def test_bm25_limit(collection_factory: CollectionFactory, limit: int) -> None:
 def test_bm25_offset(collection_factory: CollectionFactory, offset: int, expected: int) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT, tokenization=Tokenization.WORD)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     res = collection.data.insert_many(
@@ -638,7 +627,7 @@ def test_bm25_offset(collection_factory: CollectionFactory, offset: int, expecte
 def test_fetch_objects_offset(collection_factory: CollectionFactory, offset: int) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     nr_objects = 5
@@ -653,7 +642,7 @@ def test_fetch_objects_offset(collection_factory: CollectionFactory, offset: int
 def test_fetch_objects_limit(collection_factory: CollectionFactory, limit: int) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     for i in range(5):
         collection.data.insert({"Name": str(i)})
@@ -664,7 +653,7 @@ def test_fetch_objects_limit(collection_factory: CollectionFactory, limit: int) 
 def test_search_after(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     nr_objects = 10
@@ -680,7 +669,7 @@ def test_search_after(collection_factory: CollectionFactory) -> None:
 def test_auto_limit(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         inverted_index_config=Configure.inverted_index(),
     )
     for _ in range(4):
@@ -713,7 +702,7 @@ def test_query_properties(collection_factory: CollectionFactory) -> None:
             Property(name="Name", data_type=DataType.TEXT),
             Property(name="Age", data_type=DataType.INT),
         ],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection.data.insert({"Name": "rain", "Age": 1})
     collection.data.insert({"Name": "sun", "Age": 2})
@@ -739,7 +728,9 @@ def test_query_properties(collection_factory: CollectionFactory) -> None:
 def test_near_object(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
     )
     uuid_banana = collection.data.insert({"Name": "Banana"})
     collection.data.insert({"Name": "Fruit"})
@@ -765,7 +756,9 @@ def test_near_object(collection_factory: CollectionFactory) -> None:
 def test_near_object_limit(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
     )
     uuid_banana = collection.data.insert({"Name": "Banana"})
     uuid_fruit = collection.data.insert({"Name": "Fruit"})
@@ -783,7 +776,9 @@ def test_near_object_limit(collection_factory: CollectionFactory) -> None:
 def test_near_object_offset(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="Name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
     )
     uuid_banana = collection.data.insert({"Name": "Banana"})
     uuid_fruit = collection.data.insert({"Name": "Fruit"})
@@ -803,7 +798,9 @@ def test_near_object_group_by_argument(collection_factory: CollectionFactory) ->
             Property(name="Name", data_type=DataType.TEXT),
             Property(name="Count", data_type=DataType.INT),
         ],
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
     )
     uuid_banana1 = collection.data.insert({"Name": "Banana", "Count": 51})
     collection.data.insert({"Name": "Banana", "Count": 72})
@@ -823,14 +820,14 @@ def test_near_object_group_by_argument(collection_factory: CollectionFactory) ->
     assert len(ret.objects) == 4
     assert ret.objects[0].belongs_to_group == "Banana"
     assert ret.objects[1].belongs_to_group == "Banana"
-    assert ret.objects[2].belongs_to_group == "Mountain"
-    assert ret.objects[3].belongs_to_group == "car"
+    assert ret.objects[2].belongs_to_group == "car"
+    assert ret.objects[3].belongs_to_group == "Mountain"
 
 
 def test_multi_searches(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
         properties=[Property(name="name", data_type=DataType.TEXT)],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     collection.data.insert(properties={"name": "word"})
@@ -857,7 +854,7 @@ def test_multi_searches(collection_factory: CollectionFactory) -> None:
 
 def test_fetch_objects_with_limit(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[Property(name="name", data_type=DataType.TEXT)],
     )
 
@@ -870,7 +867,7 @@ def test_fetch_objects_with_limit(collection_factory: CollectionFactory) -> None
 
 def test_add_property(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[Property(name="name", data_type=DataType.TEXT)],
     )
     uuid1 = collection.data.insert({"name": "first"})
@@ -885,7 +882,9 @@ def test_add_property(collection_factory: CollectionFactory) -> None:
 
 def test_add_property_with_vectorizer(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
         properties=[Property(name="name", data_type=DataType.TEXT)],
     )
     collection.config.add_property(
@@ -925,7 +924,7 @@ def test_add_property_with_vectorizer(collection_factory: CollectionFactory) -> 
 
 def test_add_reference(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[Property(name="name", data_type=DataType.TEXT)],
     )
     uuid1 = collection.data.insert({"name": "first"})
@@ -947,7 +946,7 @@ def test_add_reference(collection_factory: CollectionFactory) -> None:
 
 def test_collection_config_get(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="name", data_type=DataType.TEXT),
             Property(name="age", data_type=DataType.INT),
@@ -960,9 +959,7 @@ def test_collection_config_get(collection_factory: CollectionFactory) -> None:
     assert config.properties[0].data_type == DataType.TEXT
     assert config.properties[1].name == "age"
     assert config.properties[1].data_type == DataType.INT
-    assert config.vectorizer is None
-    assert config.vector_config is not None
-    assert config.vector_config["default"].vectorizer.vectorizer == Vectorizers.NONE
+    assert config.vectorizer == Vectorizers.NONE
 
 
 @pytest.mark.parametrize("return_properties", [None, [], ["name"], False, True])
@@ -994,7 +991,7 @@ def test_return_properties_metadata_references_combos(
     include_vector: bool,
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="name", data_type=DataType.TEXT),
             Property(name="age", data_type=DataType.INT),
@@ -1070,7 +1067,7 @@ def test_insert_date_property(
     collection_factory: CollectionFactory, hours: int, minutes: int, sign: int
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[Property(name="date", data_type=DataType.DATE)],
     )
 
@@ -1091,7 +1088,7 @@ def test_insert_date_property(
 
 def test_exist(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     uuid1 = collection.data.insert({})
@@ -1104,7 +1101,7 @@ def test_return_list_properties(collection_factory: CollectionFactory) -> None:
     name_small = "TestReturnList"
     collection = collection_factory(
         name=name_small,
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="ints", data_type=DataType.INT_ARRAY),
             Property(name="floats", data_type=DataType.NUMBER_ARRAY),
@@ -1146,7 +1143,9 @@ def test_near_text(
     return_properties: Optional[PROPERTIES],
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
         properties=[Property(name="value", data_type=DataType.TEXT)],
     )
 
@@ -1177,7 +1176,7 @@ def test_near_text(
 
 def test_near_text_error(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
 
     with pytest.raises(ValueError):
@@ -1185,14 +1184,10 @@ def test_near_text_error(collection_factory: CollectionFactory) -> None:
 
 
 def test_near_text_group_by_argument(collection_factory: CollectionFactory) -> None:
-    dummy = collection_factory("dummy")
-    if dummy._connection._weaviate_version.is_lower_than(1, 27, 0):
-        pytest.skip(
-            'Vectors not passed to GroupByHit in servers for versions < 1.27.0. Therefore `assert "default" in ret.objects[i].vector` fails always.'
-        )
-
     collection = collection_factory(
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
         properties=[Property(name="value", data_type=DataType.TEXT)],
     )
 
@@ -1227,7 +1222,9 @@ def test_near_text_group_by_argument(collection_factory: CollectionFactory) -> N
 
 def test_near_text_limit(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
         properties=[Property(name="value", data_type=DataType.TEXT)],
     )
 
@@ -1255,7 +1252,9 @@ def test_near_text_limit(collection_factory: CollectionFactory) -> None:
 
 def test_near_text_offset(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
         properties=[Property(name="value", data_type=DataType.TEXT)],
     )
 
@@ -1303,7 +1302,7 @@ def test_near_image(
     certainty: Optional[float],
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.img2vec_neural(image_fields=["imageProp"]),
+        vectorizer_config=Configure.Vectorizer.img2vec_neural(image_fields=["imageProp"]),
         properties=[
             Property(name="imageProp", data_type=DataType.BLOB),
         ],
@@ -1360,7 +1359,7 @@ def test_near_media(
     certainty: Optional[float],
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.img2vec_neural(image_fields=["imageProp"]),
+        vectorizer_config=Configure.Vectorizer.img2vec_neural(image_fields=["imageProp"]),
         properties=[
             Property(name="imageProp", data_type=DataType.BLOB),
         ],
@@ -1408,7 +1407,7 @@ def test_return_properties_with_query_specific_typed_dict(
     collection_factory: CollectionFactory, which_case: int
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="int_", data_type=DataType.INT),
             Property(name="ints", data_type=DataType.INT_ARRAY),
@@ -1470,7 +1469,7 @@ def test_return_properties_with_general_typed_dict(collection_factory: Collectio
         ints: List[int]
 
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="int_", data_type=DataType.INT),
             Property(name="ints", data_type=DataType.INT_ARRAY),
@@ -1495,7 +1494,7 @@ def test_return_properties_with_query_specific_typed_dict_overwriting_general_ty
         int_: int
 
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="int_", data_type=DataType.INT),
             Property(name="ints", data_type=DataType.INT_ARRAY),
@@ -1516,7 +1515,7 @@ def test_return_properties_with_query_specific_typed_dict_overwriting_general_ty
 
 def test_batch_with_arrays(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="texts", data_type=DataType.TEXT_ARRAY),
             Property(name="ints", data_type=DataType.INT_ARRAY),
@@ -1582,7 +1581,7 @@ def test_sort(
     expected: List[int],
 ) -> None:
     collection = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="age", data_type=DataType.INT),
             Property(name="name", data_type=DataType.TEXT),
@@ -1608,7 +1607,7 @@ def test_sort(
 def test_optional_ref_returns(collection_factory: CollectionFactory) -> None:
     ref_collection = collection_factory(
         name="target",
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[Property(name="text", data_type=DataType.TEXT)],
     )
     uuid_to1 = ref_collection.data.insert(properties={"text": "ref text"})
@@ -1618,7 +1617,7 @@ def test_optional_ref_returns(collection_factory: CollectionFactory) -> None:
         references=[
             ReferenceProperty(name="ref", target_collection=ref_collection.name),
         ],
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection.data.insert({}, references={"ref": uuid_to1})
 
@@ -1637,7 +1636,7 @@ def test_return_properties_with_type_hint_generic(
     value: str,
 ) -> None:
     dummy = collection_factory(
-        vector_config=Configure.Vectors.self_provided(),
+        vectorizer_config=Configure.Vectorizer.none(),
         properties=[
             Property(name="name", data_type=DataType.TEXT),
         ],
@@ -1739,7 +1738,9 @@ def test_none_query_hybrid_bm25(collection_factory: CollectionFactory) -> None:
         properties=[
             Property(name="text", data_type=DataType.TEXT),
         ],
-        vector_config=Configure.Vectors.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=Configure.Vectorizer.text2vec_contextionary(
+            vectorize_collection_name=False
+        ),
     )
 
     collection.data.insert({"text": "banana"})
