+++ integration/test_collection_aggregate.py
@@ -17,12 +17,7 @@ from weaviate.collections.classes.aggregate import (
     GroupByAggregate,
     Metrics,
 )
-from weaviate.collections.classes.config import (
-    Configure,
-    DataType,
-    Property,
-    ReferenceProperty,
-)
+from weaviate.collections.classes.config import Configure, DataType, Property, ReferenceProperty
 from weaviate.collections.classes.filters import Filter, _Filters
 from weaviate.collections.classes.grpc import Move
 from weaviate.collections.classes.tenants import Tenant
