+++ integration/test_named_vectors.py
@@ -16,10 +16,7 @@ from weaviate.collections.classes.config import (
     _VectorIndexConfigHNSW,
 )
 from weaviate.collections.classes.data import DataObject
-from weaviate.collections.classes.grpc import (
-    _ListOfVectorsQuery,
-    _MultiTargetVectorJoin,
-)
+from weaviate.collections.classes.grpc import _ListOfVectorsQuery, _MultiTargetVectorJoin
 from weaviate.exceptions import WeaviateInvalidInputError, WeaviateQueryError
 from weaviate.types import INCLUDE_VECTOR
 
@@ -36,23 +33,23 @@ def test_create_named_vectors(
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="title", source_properties=["title"], vectorize_collection_name=False
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "title", source_properties=["title"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
                 name="content", source_properties=["content"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
                 name="All", vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
                 name="AllExplicit",
                 source_properties=["title", "content"],
                 vectorize_collection_name=False,
             ),
-            wvc.config.Configure.Vectors.self_provided(name="bringYourOwn"),
-            wvc.config.Configure.Vectors.self_provided(name="bringYourOwn2"),
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn"),
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn2"),
         ],
     )
 
@@ -88,11 +85,11 @@ def test_insert_many_add(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="title", source_properties=["title"], vectorize_collection_name=False
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "title", source_properties=["title"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.self_provided(name="bringYourOwn"),
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn"),
         ],
     )
 
@@ -117,7 +114,9 @@ def test_update(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=wvc.config.Configure.Vectors.self_provided(name="bringYourOwn"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn"),
+        ],
     )
 
     uuid = collection.data.insert(
@@ -146,7 +145,9 @@ def test_replace(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=wvc.config.Configure.Vectors.self_provided(name="bringYourOwn"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn"),
+        ],
     )
 
     uuid = collection.data.insert(
@@ -176,11 +177,11 @@ def test_query(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="title", source_properties=["title"], vectorize_collection_name=False
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "title", source_properties=["title"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
                 name="content", source_properties=["content"], vectorize_collection_name=False
             ),
         ],
@@ -206,11 +207,11 @@ def test_generate(openai_collection: OpenAICollection) -> None:
     if collection._connection._weaviate_version.is_lower_than(1, 24, 0):
         pytest.skip("Named vectors are not supported in versions lower than 1.24.0")
     collection = openai_collection(
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_openai(
-                name="text", source_properties=["text"], vectorize_collection_name=False
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_openai(
+                "text", source_properties=["text"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.text2vec_openai(
+            wvc.config.Configure.NamedVectors.text2vec_openai(
                 name="content", source_properties=["content"], vectorize_collection_name=False
             ),
         ],
@@ -251,11 +252,11 @@ def test_batch_add(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="content", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="title", source_properties=["title"], vectorize_collection_name=False
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "title", source_properties=["title"], vectorize_collection_name=False
             ),
-            wvc.config.Configure.Vectors.self_provided(name="bringYourOwn"),
+            wvc.config.Configure.NamedVectors.none(name="bringYourOwn"),
         ],
     )
     uuid1 = uuid.uuid4()
@@ -278,21 +279,21 @@ def test_named_vector_with_index_config(collection_factory: CollectionFactory) -
             wvc.config.Property(name="title", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="second", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="title",
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "title",
                 source_properties=["title"],
                 vectorize_collection_name=False,
                 vector_index_config=wvc.config.Configure.VectorIndex.flat(
                     distance_metric=wvc.config.VectorDistances.HAMMING,
+                    quantizer=wvc.config.Configure.VectorIndex.Quantizer.bq(rescore_limit=10),
                 ),
-                quantizer=wvc.config.Configure.VectorIndex.Quantizer.bq(rescore_limit=10),
             ),
-            wvc.config.Configure.Vectors.self_provided(
-                name="custom",
+            wvc.config.Configure.NamedVectors.none(
+                "custom",
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="default",
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "default",
                 vectorize_collection_name=False,  # needed as contextionary cant handle "_" in collection names
             ),
         ],
@@ -333,15 +334,15 @@ def test_aggregation(collection_factory: CollectionFactory) -> None:
             wvc.config.Property(name="second", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="number", data_type=wvc.config.DataType.INT),
         ],
-        # vector_config=wvc.config.Configure.Vectorizer.text2vec_contextionary(vectorize_collection_name=False),
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="first",
+        # vectorizer_config=wvc.config.Configure.Vectorizer.text2vec_contextionary(vectorize_collection_name=False),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "first",
                 source_properties=["first"],
                 vectorize_collection_name=False,
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="second",
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "second",
                 source_properties=["second"],
                 vectorize_collection_name=False,
             ),
@@ -407,14 +408,14 @@ def test_update_to_enable_quantizer_on_specific_named_vector(
             wvc.config.Property(name="first", data_type=wvc.config.DataType.TEXT),
             wvc.config.Property(name="second", data_type=wvc.config.DataType.TEXT),
         ],
-        vector_config=[
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="first",
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "first",
                 source_properties=["first"],
                 vectorize_collection_name=False,
             ),
-            wvc.config.Configure.Vectors.text2vec_contextionary(
-                name="second",
+            wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                "second",
                 source_properties=["second"],
                 vectorize_collection_name=False,
             ),
@@ -428,8 +429,8 @@ def test_update_to_enable_quantizer_on_specific_named_vector(
     assert config.vector_config["second"].vector_index_config.quantizer is None
 
     collection.config.update(
-        vector_config=[
-            wvc.config.Reconfigure.Vectors.update(
+        vectorizer_config=[
+            wvc.config.Reconfigure.NamedVectors.update(
                 name="second",
                 vector_index_config=wvc.config.Reconfigure.VectorIndex.hnsw(
                     quantizer=wvc.config.Reconfigure.VectorIndex.Quantizer.pq()
@@ -457,13 +458,13 @@ def test_update_to_enable_quantizer_on_specific_named_vector(
 #             wvc.config.Property(name="first", data_type=wvc.config.DataType.TEXT),
 #             wvc.config.Property(name="second", data_type=wvc.config.DataType.TEXT),
 #         ],
-#         vector_config=[
-#             wvc.config.Configure.Vectors.text2vec_contextionary(
+#         vectorizer_config=[
+#             wvc.config.Configure.NamedVectors.text2vec_contextionary(
 #                 "first",
 #                 source_properties=["first"],
 #                 vectorize_collection_name=False,
 #             ),
-#             wvc.config.Configure.Vectors.text2vec_contextionary(
+#             wvc.config.Configure.NamedVectors.text2vec_contextionary(
 #                 "second",
 #                 source_properties=["second"],
 #                 vectorize_collection_name=False,
@@ -482,8 +483,8 @@ def test_update_to_enable_quantizer_on_specific_named_vector(
 
 #     with pytest.raises(WeaviateInvalidInputError):
 #         collection.config.update(
-#             vector_config=[
-#                 wvc.config.ReConfigure.Vectors.update(
+#             vectorizer_config=[
+#                 wvc.config.Reconfigure.NamedVectors.update(
 #                     name="second",
 #                     vector_index_config=wvc.config.Reconfigure.VectorIndex.hnsw(
 #                         quantizer=wvc.config.Reconfigure.VectorIndex.Quantizer.bq()
@@ -496,12 +497,12 @@ def test_update_to_enable_quantizer_on_specific_named_vector(
 def test_duplicate_named_vectors(collection_factory: CollectionFactory) -> None:
     with pytest.raises(WeaviateInvalidInputError) as e:
         collection_factory(
-            vector_config=[
-                wvc.config.Configure.Vectors.text2vec_contextionary(
-                    name="title", source_properties=["title"], vectorize_collection_name=False
+            vectorizer_config=[
+                wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                    "title", source_properties=["title"], vectorize_collection_name=False
                 ),
-                wvc.config.Configure.Vectors.text2vec_contextionary(
-                    name="title", source_properties=["content"], vectorize_collection_name=False
+                wvc.config.Configure.NamedVectors.text2vec_contextionary(
+                    "title", source_properties=["content"], vectorize_collection_name=False
                 ),
             ],
         )
@@ -528,9 +529,9 @@ def test_named_vector_multi_target(
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -548,9 +549,9 @@ def test_named_vector_multi_target_vector_per_target(collection_factory: Collect
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -570,9 +571,9 @@ def test_multi_query_error_no_target_vector(collection_factory: CollectionFactor
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -581,7 +582,7 @@ def test_multi_query_error_no_target_vector(collection_factory: CollectionFactor
         with pytest.raises(WeaviateInvalidInputError):
             collection.query.near_vector([[1.0, 0.0], [1.0, 0.0, 0.0]])
         with pytest.raises(WeaviateInvalidInputError):
-            collection.query.near_vector([[[1.0, 0.0], [1.0, 0.0]], [1.0, 0.0, 0.0]])  # type: ignore
+            collection.query.near_vector([[[1.0, 0.0], [1.0, 0.0]], [1.0, 0.0, 0.0]])
     else:
         # throws an error in the server instead as implicit multi vector is understood now as using multi-vectors
         with pytest.raises(WeaviateQueryError):
@@ -617,9 +618,9 @@ def test_same_target_vector_multiple_input(
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -669,9 +670,9 @@ def test_same_target_vector_multiple_input_combinations(
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -695,9 +696,9 @@ def test_deprecated_syntax(collection_factory: CollectionFactory):
 
     collection = collection_factory(
         properties=[],
-        vector_config=[
-            wvc.config.Configure.Vectors.self_provided(name="first"),
-            wvc.config.Configure.Vectors.self_provided(name="second"),
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none("first"),
+            wvc.config.Configure.NamedVectors.none("second"),
         ],
     )
 
@@ -706,7 +707,7 @@ def test_deprecated_syntax(collection_factory: CollectionFactory):
 
     with pytest.raises(WeaviateInvalidInputError) as e:
         collection.query.near_vector(
-            [[0.0, 1.0], [[1.0, 0.0, 0.0], [0.0, 0.0, 1.0]]],  # # type: ignore
+            [[0.0, 1.0], [[1.0, 0.0, 0.0], [0.0, 0.0, 1.0]]],
             target_vector=["first", "second", "second"],
             return_metadata=wvc.query.MetadataQuery.full(),
         )
@@ -726,7 +727,8 @@ def test_deprecated_syntax(collection_factory: CollectionFactory):
     [
         (False, {}),
         (["bringYourOwn1"], {"bringYourOwn1": [0, 1, 2]}),
-        (True, {"bringYourOwn1": [0, 1, 2], "bringYourOwn2": [3, 4, 5]}),
+        # TODO: to be uncommented when https://github.com/weaviate/weaviate/issues/6279 is resolved
+        # (True, {"bringYourOwn1": [0, 1, 2], "bringYourOwn2": [3, 4, 5]})
     ],
 )
 def test_include_vector_on_references(
@@ -734,10 +736,8 @@ def test_include_vector_on_references(
 ) -> None:
     """Test include vector on reference."""
     dummy = collection_factory()
-    if dummy._connection._weaviate_version.is_lower_than(1, 26, 0):
-        pytest.skip(
-            "https://github.com/weaviate/weaviate/issues/6279 was resolved in >1.26.0 only."
-        )
+    if dummy._connection._weaviate_version.is_lower_than(1, 24, 0):
+        pytest.skip("Named vectorizers are only supported in Weaviate v1.24.0 and higher.")
 
     ref_collection = collection_factory(
         name="Target",
@@ -777,14 +777,16 @@ def test_colbert_vectors_byov(collection_factory: CollectionFactory) -> None:
                 data_type=wvc.config.DataType.TEXT,
             )
         ],
-        vector_config=[
-            wvc.config.Configure.MultiVectors.self_provided(
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.none(
                 name="colbert",
-                multi_vector_config=wvc.config.Configure.VectorIndex.MultiVector.multi_vector(
-                    aggregation=wvc.config.MultiVectorAggregation.MAX_SIM
+                vector_index_config=wvc.config.Configure.VectorIndex.hnsw(
+                    multi_vector=wvc.config.Configure.VectorIndex.MultiVector.multi_vector(
+                        aggregation=wvc.config.MultiVectorAggregation.MAX_SIM
+                    )
                 ),
             ),
-            wvc.config.Configure.Vectors.self_provided(
+            wvc.config.Configure.NamedVectors.none(
                 name="regular",
             ),
         ],
@@ -878,8 +880,8 @@ def test_colbert_vectors_jinaai(collection_factory: CollectionFactory) -> None:
                 data_type=wvc.config.DataType.TEXT,
             )
         ],
-        vector_config=[
-            wvc.config.Configure.MultiVectors.text2vec_jinaai(
+        vectorizer_config=[
+            wvc.config.Configure.NamedVectors.text2colbert_jinaai(
                 name="colbert",
             )
         ],
