+++ test/collection/test_config.py
@@ -15,20 +15,11 @@ from weaviate.collections.classes.config import (
     _VectorizerConfigCreate,
 )
 from weaviate.collections.classes.config_named_vectors import _NamedVectorConfigCreate
-from weaviate.collections.classes.config_vectorizers import (
-    Multi2VecField,
-    VectorDistances,
-)
+from weaviate.collections.classes.config_vectorizers import Multi2VecField, VectorDistances
 
 DEFAULTS = {
-    "vectorConfig": {
-        "default": {
-            "vectorIndexType": "hnsw",
-            "vectorizer": {
-                "none": {},
-            },
-        }
-    }
+    "vectorizer": "none",
+    "vectorIndexType": "hnsw",
 }
 
 
@@ -683,7 +674,7 @@ def test_config_with_default_vectorizer(
 ) -> None:
     config = _CollectionConfigCreate(name="test", vectorizer_config=vectorizer_config)
     assert config._to_dict() == {
-        "vectorIndexType": "hnsw",
+        **DEFAULTS,
         "vectorizer": vectorizer_config.vectorizer.value,
         "class": "Test",
         "moduleConfig": expected,
@@ -792,7 +783,7 @@ def test_config_with_vectorizer_and_properties(
         name="test", properties=properties, vectorizer_config=vectorizer_config
     )
     assert config._to_dict() == {
-        "vectorIndexType": "hnsw",
+        **DEFAULTS,
         "vectorizer": vectorizer_config.vectorizer.value,
         "class": "Test",
         "properties": expected_props,
@@ -1066,6 +1057,7 @@ def test_config_with_generative(
     config = _CollectionConfigCreate(name="test", generative_config=generative_config)
     assert config._to_dict() == {
         **DEFAULTS,
+        "vectorizer": "none",
         "class": "Test",
         "moduleConfig": expected_mc,
     }
@@ -1145,6 +1137,7 @@ def test_config_with_reranker(
     config = _CollectionConfigCreate(name="test", reranker_config=reranker_config)
     assert config._to_dict() == {
         **DEFAULTS,
+        "vectorizer": "none",
         "class": "Test",
         "moduleConfig": expected_mc,
     }
