+++ weaviate/classes/config.py
@@ -3,6 +3,7 @@ from weaviate.collections.classes.config import (
     ConsistencyLevel,
     DataType,
     GenerativeSearches,
+    MultiVectorAggregation,
     PQEncoderDistribution,
     PQEncoderType,
     Property,
@@ -14,10 +15,7 @@ from weaviate.collections.classes.config import (
     Tokenization,
     VectorDistances,
 )
-from weaviate.collections.classes.config_vector_index import (
-    MultiVectorAggregation,
-    VectorFilterStrategy,
-)
+from weaviate.collections.classes.config_vector_index import VectorFilterStrategy
 from weaviate.collections.classes.config_vectorizers import Multi2VecField, Vectorizers
 from weaviate.connect.integrations import Integrations
