+++ weaviate/collections/classes/config.py
@@ -32,19 +32,19 @@ from weaviate.collections.classes.config_named_vectors import (
     _NamedVectorsUpdate,
 )
 from weaviate.collections.classes.config_vector_index import (
-    PQEncoderDistribution,
-    PQEncoderType,
     VectorFilterStrategy,
-    _BQConfigUpdate,
-    _PQConfigUpdate,
-    _PQEncoderConfigUpdate,
-    _RQConfigUpdate,
-    _SQConfigUpdate,
-    _VectorIndex,
+    _EncodingConfigCreate,
+    _MultiVectorConfigCreate,
+    _MuveraConfigCreate,
+    _QuantizerConfigCreate,
     _VectorIndexConfigCreate,
+    _VectorIndexConfigDynamicCreate,
     _VectorIndexConfigDynamicUpdate,
+    _VectorIndexConfigFlatCreate,
     _VectorIndexConfigFlatUpdate,
+    _VectorIndexConfigHNSWCreate,
     _VectorIndexConfigHNSWUpdate,
+    _VectorIndexConfigSkipCreate,
     _VectorIndexConfigUpdate,
 )
 from weaviate.collections.classes.config_vector_index import (
@@ -60,13 +60,6 @@ from weaviate.collections.classes.config_vectorizers import (
 )
 from weaviate.collections.classes.config_vectorizers import (
     Vectorizers as VectorizersAlias,
-)
-from weaviate.collections.classes.config_vectors import (
-    _MultiVectors,
-    _VectorConfigCreate,
-    _VectorConfigUpdate,
-    _Vectors,
-    _VectorsUpdate,
 )
 from weaviate.exceptions import WeaviateInvalidInputError
 from weaviate.str_enum import BaseEnum
@@ -245,6 +238,140 @@ class ReplicationDeletionStrategy(str, BaseEnum):
     DELETE_ON_CONFLICT = "DeleteOnConflict"
     NO_AUTOMATED_RESOLUTION = "NoAutomatedResolution"
     TIME_BASED_RESOLUTION = "TimeBasedResolution"
+
+
+class PQEncoderType(str, BaseEnum):
+    """Type of the PQ encoder.
+
+    Attributes:
+        KMEANS: K-means encoder.
+        TILE: Tile encoder.
+    """
+
+    KMEANS = "kmeans"
+    TILE = "tile"
+
+
+class PQEncoderDistribution(str, BaseEnum):
+    """Distribution of the PQ encoder.
+
+    Attributes:
+        LOG_NORMAL: Log-normal distribution.
+        NORMAL: Normal distribution.
+    """
+
+    LOG_NORMAL = "log-normal"
+    NORMAL = "normal"
+
+
+class MultiVectorAggregation(str, BaseEnum):
+    """Aggregation type to use for multivector indices.
+
+    Attributes:
+        MAX_SIM: Maximum similarity.
+    """
+
+    MAX_SIM = "maxSim"
+
+
+class _PQEncoderConfigCreate(_ConfigCreateModel):
+    type_: Optional[PQEncoderType] = Field(serialization_alias="type")
+    distribution: Optional[PQEncoderDistribution]
+
+
+class _PQEncoderConfigUpdate(_ConfigUpdateModel):
+    type_: Optional[PQEncoderType]
+    distribution: Optional[PQEncoderDistribution]
+
+    def merge_with_existing(self, schema: Dict[str, Any]) -> Dict[str, Any]:
+        """Must be done manually since Pydantic does not work well with type and type_.
+
+        Errors shadowing type occur if we want to use type as a field name.
+        """
+        if self.type_ is not None:
+            schema["type"] = str(self.type_.value)
+        if self.distribution is not None:
+            schema["distribution"] = str(self.distribution.value)
+        return schema
+
+
+class _PQConfigCreate(_QuantizerConfigCreate):
+    bitCompression: Optional[bool] = Field(default=None)
+    centroids: Optional[int]
+    encoder: _PQEncoderConfigCreate
+    segments: Optional[int]
+    trainingLimit: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "pq"
+
+
+class _BQConfigCreate(_QuantizerConfigCreate):
+    cache: Optional[bool]
+    rescoreLimit: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "bq"
+
+
+class _SQConfigCreate(_QuantizerConfigCreate):
+    cache: Optional[bool]
+    rescoreLimit: Optional[int]
+    trainingLimit: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "sq"
+
+
+class _RQConfigCreate(_QuantizerConfigCreate):
+    bits: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "rq"
+
+
+class _PQConfigUpdate(_QuantizerConfigUpdate):
+    bitCompression: Optional[bool] = Field(default=None)
+    centroids: Optional[int]
+    enabled: Optional[bool]
+    segments: Optional[int]
+    trainingLimit: Optional[int]
+    encoder: Optional[_PQEncoderConfigUpdate]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "pq"
+
+
+class _BQConfigUpdate(_QuantizerConfigUpdate):
+    enabled: Optional[bool]
+    rescoreLimit: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "bq"
+
+
+class _SQConfigUpdate(_QuantizerConfigUpdate):
+    enabled: Optional[bool]
+    rescoreLimit: Optional[int]
+    trainingLimit: Optional[int]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "sq"
+
+
+class _RQConfigUpdate(_QuantizerConfigUpdate):
+    enabled: Optional[bool]
+
+    @staticmethod
+    def quantizer_name() -> str:
+        return "rq"
 
 
 class _ShardingConfigCreate(_ConfigCreateModel):
@@ -1101,32 +1228,11 @@ class _CollectionConfigUpdate(_ConfigUpdateModel):
     vectorizerConfig: Optional[Union[_VectorIndexConfigUpdate, List[_NamedVectorConfigUpdate]]] = (
         Field(default=None, alias="vectorizer_config")
     )
-    vectorConfig: Optional[Union[_VectorConfigUpdate, List[_VectorConfigUpdate]]] = Field(
-        default=None, alias="vector_config"
-    )
     multiTenancyConfig: Optional[_MultiTenancyConfigUpdate] = Field(
         default=None, alias="multi_tenancy_config"
     )
     generativeConfig: Optional[_GenerativeProvider] = Field(default=None, alias="generative_config")
     rerankerConfig: Optional[_RerankerProvider] = Field(default=None, alias="reranker_config")
-
-    @field_validator("vectorConfig", mode="before")
-    def mutual_exclusivity(
-        cls,
-        v: Optional[Union[_VectorConfigUpdate, List[_VectorConfigUpdate]]],
-        info: ValidationInfo,
-    ):
-        if v is None:
-            return v
-        if info.data["vectorizerConfig"] is not None:
-            raise ValueError(
-                "Cannot specify vectorizerConfig when also specifying vectorConfig. Please use one or the other."
-            )
-        if info.data["vectorIndexConfig"] is not None:
-            raise ValueError(
-                "Cannot specify vectorIndexConfig when also specifying vectorConfig. Please use one or the other."
-            )
-        return v
 
     def __check_quantizers(
         self,
@@ -1252,29 +1358,6 @@ class _CollectionConfigUpdate(_ConfigUpdateModel):
                     schema["vectorConfig"][vc.name]["vectorIndexType"] = (
                         vc.vectorIndexConfig.vector_index_type()
                     )
-        if self.vectorConfig is not None:
-            vcs = (
-                [self.vectorConfig]
-                if isinstance(self.vectorConfig, _VectorConfigUpdate)
-                else self.vectorConfig
-            )
-            for vc in vcs:
-                if vc.name not in schema["vectorConfig"]:
-                    raise WeaviateInvalidInputError(
-                        f"Vector config with name {vc.name} does not exist in the existing vector config"
-                    )
-                self.__check_quantizers(
-                    vc.vectorIndexConfig.quantizer,
-                    schema["vectorConfig"][vc.name]["vectorIndexConfig"],
-                )
-                schema["vectorConfig"][vc.name]["vectorIndexConfig"] = (
-                    vc.vectorIndexConfig.merge_with_existing(
-                        schema["vectorConfig"][vc.name]["vectorIndexConfig"]
-                    )
-                )
-                schema["vectorConfig"][vc.name]["vectorIndexType"] = (
-                    vc.vectorIndexConfig.vector_index_type()
-                )
         return schema
 
     @staticmethod
@@ -1889,13 +1972,8 @@ class _CollectionConfigCreate(_ConfigCreateModel):
     vectorIndexConfig: Optional[_VectorIndexConfigCreate] = Field(
         default=None, alias="vector_index_config"
     )
-    vectorizerConfig: Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate], None] = Field(
-        default=None, alias="vectorizer_config"
-    )
-    vectorConfig: Union[_VectorConfigCreate, List[_VectorConfigCreate], None] = Field(
-        default=None,
-        alias="vector_config",
-        validate_default=True,
+    vectorizerConfig: Optional[Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]]] = (
+        Field(default=_Vectorizer.none(), alias="vectorizer_config")
     )
     generativeSearch: Optional[_GenerativeProvider] = Field(default=None, alias="generative_config")
     rerankerConfig: Optional[_RerankerProvider] = Field(default=None, alias="reranker_config")
@@ -1903,31 +1981,18 @@ class _CollectionConfigCreate(_ConfigCreateModel):
     def model_post_init(self, __context: Any) -> None:
         self.name = _capitalize_first_letter(self.name)
 
-    @field_validator("vectorizerConfig", "vectorConfig", mode="after")
+    @field_validator("vectorizerConfig", mode="after")
     @classmethod
     def validate_vector_names(
         cls,
-        v: Union[_VectorizerConfigCreate, _NamedVectorConfigCreate, List[_NamedVectorConfigCreate]],
+        v: Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]],
         info: ValidationInfo,
-    ) -> Union[_VectorizerConfigCreate, _NamedVectorConfigCreate, List[_NamedVectorConfigCreate]]:
+    ) -> Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]]:
         if isinstance(v, list):
             names = [vc.name for vc in v]
             if len(names) != len(set(names)):
                 dups = {name for name in names if names.count(name) > 1}
                 raise ValueError(f"Vector config names must be unique. Found duplicates: {dups}")
-        return v
-
-    @field_validator("vectorConfig", mode="after")
-    @classmethod
-    def inject_vector_config_none(
-        cls,
-        v: Union[_VectorConfigCreate, List[_VectorConfigCreate], None],
-        info: ValidationInfo,
-    ) -> Union[_VectorConfigCreate, List[_VectorConfigCreate], None]:
-        if v is None and info.data["vectorizerConfig"] is None:
-            return _VectorConfigCreate(
-                name="default", vectorizer=_VectorizerConfigCreate(vectorizer=Vectorizers.NONE)
-            )
         return v
 
     @staticmethod
@@ -1959,8 +2024,6 @@ class _CollectionConfigCreate(_ConfigCreateModel):
             elif isinstance(val, _VectorIndexConfigCreate):
                 ret_dict["vectorIndexType"] = val.vector_index_type().value
                 ret_dict[cls_field] = val._to_dict()
-            elif isinstance(val, _VectorConfigCreate):
-                ret_dict["vectorConfig"] = {val.name or "default": val._to_dict()}
             elif (
                 isinstance(val, list)
                 and len(val) > 0
@@ -1968,19 +2031,7 @@ class _CollectionConfigCreate(_ConfigCreateModel):
             ):
                 val = cast(List[_NamedVectorConfigCreate], val)
                 ret_dict["vectorConfig"] = {item.name: item._to_dict() for item in val}
-            elif (
-                isinstance(val, list)
-                and len(val) > 0
-                and all(isinstance(item, _VectorConfigCreate) for item in val)
-            ):
-                val = cast(List[_VectorConfigCreate], val)
-                ret_dict["vectorConfig"] = {}
-                for item in val:
-                    if item.name is None:
-                        raise WeaviateInvalidInputError(
-                            "Vector config name must be set when specifying multiple vectors"
-                        )
-                    ret_dict["vectorConfig"][item.name] = item._to_dict()
+
             else:
                 assert isinstance(val, _ConfigCreateModel)
                 ret_dict[cls_field] = val._to_dict()
@@ -2027,6 +2078,212 @@ class _CollectionConfigCreate(_ConfigCreateModel):
         ret_dict["properties"] = existing_props
 
 
+class _VectorIndexMultivectorEncoding:
+    @staticmethod
+    def muvera(
+        ksim: Optional[int] = None,
+        dprojections: Optional[int] = None,
+        repetitions: Optional[int] = None,
+    ) -> _EncodingConfigCreate:
+        return _MuveraConfigCreate(
+            enabled=True,
+            ksim=ksim,
+            dprojections=dprojections,
+            repetitions=repetitions,
+        )
+
+
+class _VectorIndexMultiVector:
+    Encoding = _VectorIndexMultivectorEncoding
+
+    @staticmethod
+    def multi_vector(
+        encoding: Optional[_EncodingConfigCreate] = None,
+        aggregation: Optional[MultiVectorAggregation] = None,
+    ) -> _MultiVectorConfigCreate:
+        return _MultiVectorConfigCreate(
+            encoding=encoding if encoding is not None else None,
+            aggregation=aggregation.value if aggregation is not None else None,
+        )
+
+
+class _VectorIndexQuantizer:
+    @staticmethod
+    def pq(
+        bit_compression: Optional[bool] = None,
+        centroids: Optional[int] = None,
+        encoder_distribution: Optional[PQEncoderDistribution] = None,
+        encoder_type: Optional[PQEncoderType] = None,
+        segments: Optional[int] = None,
+        training_limit: Optional[int] = None,
+    ) -> _PQConfigCreate:
+        """Create a `_PQConfigCreate` object to be used when defining the product quantization (PQ) configuration of Weaviate.
+
+        Use this method when defining the `quantizer` argument in the `vector_index` configuration.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/concepts/vector-index#hnsw-with-compression) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        if bit_compression is not None:
+            _Warnings.bit_compression_in_pq_config()
+        return _PQConfigCreate(
+            centroids=centroids,
+            segments=segments,
+            trainingLimit=training_limit,
+            encoder=_PQEncoderConfigCreate(type_=encoder_type, distribution=encoder_distribution),
+        )
+
+    @staticmethod
+    def bq(
+        cache: Optional[bool] = None,
+        rescore_limit: Optional[int] = None,
+    ) -> _BQConfigCreate:
+        """Create a `_BQConfigCreate` object to be used when defining the binary quantization (BQ) configuration of Weaviate.
+
+        Use this method when defining the `quantizer` argument in the `vector_index` configuration. Note that the arguments have no effect for HNSW.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/concepts/vector-index#binary-quantization) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _BQConfigCreate(
+            cache=cache,
+            rescoreLimit=rescore_limit,
+        )
+
+    @staticmethod
+    def sq(
+        cache: Optional[bool] = None,
+        rescore_limit: Optional[int] = None,
+        training_limit: Optional[int] = None,
+    ) -> _SQConfigCreate:
+        """Create a `_SQConfigCreate` object to be used when defining the scalar quantization (SQ) configuration of Weaviate.
+
+        Use this method when defining the `quantizer` argument in the `vector_index` configuration. Note that the arguments have no effect for HNSW.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/concepts/vector-index#binary-quantization) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _SQConfigCreate(
+            cache=cache,
+            rescoreLimit=rescore_limit,
+            trainingLimit=training_limit,
+        )
+
+    @staticmethod
+    def rq(
+        bits: Optional[int] = None,
+    ) -> _RQConfigCreate:
+        """Create a `_RQConfigCreate` object to be used when defining the Rotational quantization (RQ) configuration of Weaviate.
+
+        Use this method when defining the `quantizer` argument in the `vector_index` configuration. Note that the arguments have no effect for HNSW.
+
+        Arguments:
+            See [the docs](https://weaviate.io/developers/weaviate/concepts/vector-index) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _RQConfigCreate(
+            bits=bits,
+        )
+
+
+class _VectorIndex:
+    MultiVector = _VectorIndexMultiVector
+    Quantizer = _VectorIndexQuantizer
+
+    @staticmethod
+    def none() -> _VectorIndexConfigSkipCreate:
+        """Create a `_VectorIndexConfigSkipCreate` object to be used when configuring Weaviate to not index your vectors.
+
+        Use this method when defining the `vector_index_config` argument in `collections.create()`.
+        """
+        return _VectorIndexConfigSkipCreate(
+            distance=None,
+            quantizer=None,
+            multivector=None,
+        )
+
+    @staticmethod
+    def hnsw(
+        cleanup_interval_seconds: Optional[int] = None,
+        distance_metric: Optional[VectorDistances] = None,
+        dynamic_ef_factor: Optional[int] = None,
+        dynamic_ef_max: Optional[int] = None,
+        dynamic_ef_min: Optional[int] = None,
+        ef: Optional[int] = None,
+        ef_construction: Optional[int] = None,
+        filter_strategy: Optional[VectorFilterStrategy] = None,
+        flat_search_cutoff: Optional[int] = None,
+        max_connections: Optional[int] = None,
+        vector_cache_max_objects: Optional[int] = None,
+        quantizer: Optional[_QuantizerConfigCreate] = None,
+        multi_vector: Optional[_MultiVectorConfigCreate] = None,
+    ) -> _VectorIndexConfigHNSWCreate:
+        """Create a `_VectorIndexConfigHNSWCreate` object to be used when defining the HNSW vector index configuration of Weaviate.
+
+        Use this method when defining the `vector_index_config` argument in `collections.create()`.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/configuration/indexes#how-to-configure-hnsw) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _VectorIndexConfigHNSWCreate(
+            cleanupIntervalSeconds=cleanup_interval_seconds,
+            distance=distance_metric,
+            dynamicEfMin=dynamic_ef_min,
+            dynamicEfMax=dynamic_ef_max,
+            dynamicEfFactor=dynamic_ef_factor,
+            efConstruction=ef_construction,
+            ef=ef,
+            filterStrategy=filter_strategy,
+            flatSearchCutoff=flat_search_cutoff,
+            maxConnections=max_connections,
+            vectorCacheMaxObjects=vector_cache_max_objects,
+            quantizer=quantizer,
+            multivector=multi_vector,
+        )
+
+    @staticmethod
+    def flat(
+        distance_metric: Optional[VectorDistances] = None,
+        vector_cache_max_objects: Optional[int] = None,
+        quantizer: Optional[_BQConfigCreate] = None,
+    ) -> _VectorIndexConfigFlatCreate:
+        """Create a `_VectorIndexConfigFlatCreate` object to be used when defining the FLAT vector index configuration of Weaviate.
+
+        Use this method when defining the `vector_index_config` argument in `collections.create()`.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/configuration/indexes#how-to-configure-hnsw) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _VectorIndexConfigFlatCreate(
+            distance=distance_metric,
+            vectorCacheMaxObjects=vector_cache_max_objects,
+            quantizer=quantizer,
+            multivector=None,
+        )
+
+    @staticmethod
+    def dynamic(
+        distance_metric: Optional[VectorDistances] = None,
+        threshold: Optional[int] = None,
+        hnsw: Optional[_VectorIndexConfigHNSWCreate] = None,
+        flat: Optional[_VectorIndexConfigFlatCreate] = None,
+    ) -> _VectorIndexConfigDynamicCreate:
+        """Create a `_VectorIndexConfigDynamicCreate` object to be used when defining the DYNAMIC vector index configuration of Weaviate.
+
+        Use this method when defining the `vector_index_config` argument in `collections.create()`.
+
+        Args:
+            See [the docs](https://weaviate.io/developers/weaviate/configuration/indexes#how-to-configure-hnsw) for a more detailed view!
+        """  # noqa: D417 (missing argument descriptions in the docstring)
+        return _VectorIndexConfigDynamicCreate(
+            distance=distance_metric,
+            threshold=threshold,
+            hnsw=hnsw,
+            flat=flat,
+            quantizer=None,
+            multivector=None,
+        )
+
+
 class Configure:
     """Use this factory class to generate the correct object for use when using the `collections.create()` method. E.g., `.multi_tenancy()` will return a `MultiTenancyConfigCreate` object to be used in the `multi_tenancy_config` argument.
 
@@ -2039,8 +2296,6 @@ class Configure:
     Vectorizer = _Vectorizer
     VectorIndex = _VectorIndex
     NamedVectors = _NamedVectors
-    Vectors = _Vectors
-    MultiVectors = _MultiVectors
 
     @staticmethod
     def inverted_index(
@@ -2312,7 +2567,6 @@ class Reconfigure:
     """
 
     NamedVectors = _NamedVectorsUpdate
-    Vectors = _VectorsUpdate
     VectorIndex = _VectorIndexUpdate
     Generative = _Generative  # config is the same for create and update
     Reranker = _Reranker  # config is the same for create and update
