+++ weaviate/collections/collections/async_.pyi
@@ -1,7 +1,5 @@
 from typing import Dict, List, Literal, Optional, Sequence, Type, Union, overload
 
-from typing_extensions import deprecated
-
 from weaviate.collections.classes.config import (
     CollectionConfig,
     CollectionConfigSimple,
@@ -14,7 +12,6 @@ from weaviate.collections.classes.config import (
     _ReplicationConfigCreate,
     _RerankerProvider,
     _ShardingConfigCreate,
-    _VectorConfigCreate,
     _VectorIndexConfigCreate,
     _VectorizerConfigCreate,
 )
@@ -27,57 +24,6 @@ from weaviate.collections.collections.base import _CollectionsBase
 from weaviate.connect.v4 import ConnectionAsync
 
 class _CollectionsAsync(_CollectionsBase[ConnectionAsync]):
-    @overload
-    @deprecated(
-        'Using the "vector_index_config" argument is deprecated. Instead, define the vector index for each specific vectorizer supplied to the "vectorizers_config" argument.'
-    )
-    async def create(
-        self,
-        name: str,
-        *,
-        description: Optional[str] = None,
-        generative_config: Optional[_GenerativeProvider] = None,
-        inverted_index_config: Optional[_InvertedIndexConfigCreate] = None,
-        multi_tenancy_config: Optional[_MultiTenancyConfigCreate] = None,
-        properties: Optional[Sequence[Property]] = None,
-        references: Optional[List[_ReferencePropertyBase]] = None,
-        replication_config: Optional[_ReplicationConfigCreate] = None,
-        reranker_config: Optional[_RerankerProvider] = None,
-        sharding_config: Optional[_ShardingConfigCreate] = None,
-        vector_index_config: _VectorIndexConfigCreate,
-        vectorizer_config: Optional[
-            Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]]
-        ] = None,
-        vector_config: Optional[Union[_VectorConfigCreate, List[_VectorConfigCreate]]] = None,
-        data_model_properties: Optional[Type[Properties]] = None,
-        data_model_references: Optional[Type[References]] = None,
-        skip_argument_validation: bool = False,
-    ) -> CollectionAsync[Properties, References]: ...
-    @overload
-    @deprecated(
-        'Using the "vectorizer_config" argument is deprecated. Instead, use the "vectorizers_config" argument.'
-    )
-    async def create(
-        self,
-        name: str,
-        *,
-        description: Optional[str] = None,
-        generative_config: Optional[_GenerativeProvider] = None,
-        inverted_index_config: Optional[_InvertedIndexConfigCreate] = None,
-        multi_tenancy_config: Optional[_MultiTenancyConfigCreate] = None,
-        properties: Optional[Sequence[Property]] = None,
-        references: Optional[List[_ReferencePropertyBase]] = None,
-        replication_config: Optional[_ReplicationConfigCreate] = None,
-        reranker_config: Optional[_RerankerProvider] = None,
-        sharding_config: Optional[_ShardingConfigCreate] = None,
-        vector_index_config: Optional[_VectorIndexConfigCreate] = None,
-        vectorizer_config: Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]],
-        vector_config: Optional[Union[_VectorConfigCreate, List[_VectorConfigCreate]]] = None,
-        data_model_properties: Optional[Type[Properties]] = None,
-        data_model_references: Optional[Type[References]] = None,
-        skip_argument_validation: bool = False,
-    ) -> CollectionAsync[Properties, References]: ...
-    @overload
     async def create(
         self,
         name: str,
@@ -95,7 +41,6 @@ class _CollectionsAsync(_CollectionsBase[ConnectionAsync]):
         vectorizer_config: Optional[
             Union[_VectorizerConfigCreate, List[_NamedVectorConfigCreate]]
         ] = None,
-        vector_config: Optional[Union[_VectorConfigCreate, List[_VectorConfigCreate]]] = None,
         data_model_properties: Optional[Type[Properties]] = None,
         data_model_references: Optional[Type[References]] = None,
         skip_argument_validation: bool = False,
