+++ weaviate/collections/config/async_.pyi
@@ -15,7 +15,6 @@ from weaviate.collections.classes.config import (
     _ReferencePropertyMultiTarget,
     _ReplicationConfigUpdate,
     _RerankerProvider,
-    _VectorConfigUpdate,
     _VectorIndexConfigFlatUpdate,
     _VectorIndexConfigHNSWUpdate,
 )
@@ -52,7 +51,6 @@ class _ConfigCollectionAsync(_ConfigCollectionExecutor[ConnectionAsync]):
                 List[_NamedVectorConfigUpdate],
             ]
         ] = None,
-        vector_config: Optional[Union[_VectorConfigUpdate, List[_VectorConfigUpdate]]] = None,
         generative_config: Optional[_GenerativeProvider] = None,
         reranker_config: Optional[_RerankerProvider] = None,
     ) -> None: ...
