+++ weaviate/collections/config/sync.pyi
@@ -15,7 +15,6 @@ from weaviate.collections.classes.config import (
     _ReferencePropertyMultiTarget,
     _ReplicationConfigUpdate,
     _RerankerProvider,
-    _VectorConfigUpdate,
     _VectorIndexConfigFlatUpdate,
     _VectorIndexConfigHNSWUpdate,
 )
@@ -50,7 +49,6 @@ class _ConfigCollection(_ConfigCollectionExecutor[ConnectionSync]):
                 List[_NamedVectorConfigUpdate],
             ]
         ] = None,
-        vector_config: Optional[Union[_VectorConfigUpdate, List[_VectorConfigUpdate]]] = None,
         generative_config: Optional[_GenerativeProvider] = None,
         reranker_config: Optional[_RerankerProvider] = None,
     ) -> None: ...
