+++ weaviate/collections/queries/base_executor.py
@@ -156,8 +156,7 @@ class _BaseExecutor(Generic[ConnectionType]):
             return {}
 
         if len(add_props.vector_bytes) > 0:
-            vec = _ByteOps.decode_float32s(add_props.vector_bytes)
-            return {"default": vec}
+            return {"default": _ByteOps.decode_float32s(add_props.vector_bytes)}
 
         vecs: Dict[str, Union[List[float], List[List[float]]]] = {}
         for vec in add_props.vectors:
